pyproject.toml
[tool.poetry]
name = "iberifier"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10, <3.11"
preprocessingtweet = "^0.1.1"
pymongo = "^4.2.0"
requests = "^2.28.1"
PyYAML = "^6.0"
google-api-python-client = "^2.64.0"
searchtweets-v2 = "^1.1.1"
tenacity = "^8.1.0"
bs4 = "^0.0.1"
pip = "^22.3"
fasttext-wheel = "^0.9.2"
wget = "^3.2"
langdetect = "^1.0.9"
langid = "^1.1.6"
transformers = "^4.23.1"
lxml = "^4.9.1"
tensorflow = "^2.10.0"
# torch = { version = "1.12.1", source="torch"}
torch = "^1.13"


# torch = { url = "https://download.pytorch.org/whl/cpu/torch-1.9.1%2Bcpu-cp39-cp39-linux_x86_64.whl"}
# torchaudio = { url = "https://download.pytorch.org/whl/cu116/torchaudio-0.12.1%2Bcu116-cp310-cp310-win_amd64.whl"}
# torchvision = { url = "https://download.pytorch.org/whl/cu116/torchvision-0.13.1%2Bcu116-cp310-cp310-win_amd64.whl"}
jupyterlab = "^3.5.3"
pandas = "^1.5.3"
seaborn = "^0.12.2"
tikzplotlib = "^0.10.1"
sklearn = "^0.0.post4"
networkx = "^3.1"
tfcausalimpact = "^0.0.13"
neuralprophet = "^0.6.0"
statsmodels = "^0.14.0"
pymer4 = "^0.8.0"
jupyterlab-lsp = "^4.1.0"


[tool.poetry.dev-dependencies]
pyls-isort = "^0.2.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# [[tool.poetry.source]]
# name = "torch"
# url = "https://download.pytorch.org/whl/cu116"
# secondary = true
